+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added keyboard navigation (Space + W/A/S/D) for moving within the zoomable image container.
 - Added support for Python 3.14.
 - Added `ImageSample.add_annotations` method, as well as `VideoSample.add_annotations`.
+- Added `SampleMetadata.update` method.
 
 ### Changed